# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/formations": {
      "get": {
        "operationId": "FormationsController_findAllPublic",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PSC1",
                "PSE1",
                "PSE2",
                "BNSSA",
                "SSA",
                "SST",
                "BSB"
              ],
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2023-12",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Récupérer toutes les formations disponibles (accès public)",
        "tags": [
          "formations"
        ]
      }
    },
    "/api/formations/admin": {
      "get": {
        "operationId": "FormationsController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PSC1",
                "PSE1",
                "PSE2",
                "BNSSA",
                "SSA",
                "SST",
                "BSB"
              ],
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2023-12",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PLANIFIEE",
                "EN_COURS",
                "TERMINEE",
                "ANNULEE"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer toutes les formations (admin)",
        "tags": [
          "formations"
        ]
      },
      "post": {
        "operationId": "FormationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFormationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créer une nouvelle formation (admin)",
        "tags": [
          "formations"
        ]
      }
    },
    "/api/formations/admin/{id}": {
      "get": {
        "operationId": "FormationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer une formation par son ID (admin)",
        "tags": [
          "formations"
        ]
      },
      "put": {
        "operationId": "FormationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une formation (admin)",
        "tags": [
          "formations"
        ]
      },
      "delete": {
        "operationId": "FormationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une formation (admin)",
        "tags": [
          "formations"
        ]
      }
    },
    "/api/formations/admin/{id}/status": {
      "put": {
        "operationId": "FormationsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFormationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour le statut d'une formation (admin)",
        "tags": [
          "formations"
        ]
      }
    },
...
    "/api/gallery": {
      "get": {
        "operationId": "GalleryController_findAll",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "enum": [
                "formations",
                "activites",
                "evenements"
              ],
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Mode de récupération (all = par catégorie, random = images aléatoires)",
            "schema": {
              "enum": [
                "all",
                "random"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des images récupérée avec succès"
          }
        },
        "summary": "Récupérer toutes les images (accès public)",
        "tags": [
          "gallery"
        ]
      }
    },
    "/api/gallery/admin": {
      "post": {
        "operationId": "GalleryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "alt": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "formations",
                      "activites",
                      "evenements"
                    ]
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image créée avec succès"
          },
          "400": {
            "description": "Données invalides ou fichier manquant"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Télécharger une nouvelle image (admin)",
        "tags": [
          "gallery"
        ]
      },
      "get": {
        "operationId": "GalleryController_findAllAdmin",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtrer par catégorie",
            "schema": {
              "enum": [
                "formations",
                "activites",
                "evenements"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des images récupérée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer toutes les images (admin)",
        "tags": [
          "gallery"
        ]
      }
    },
    "/api/gallery/admin/{id}": {
      "get": {
        "operationId": "GalleryController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'image",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image récupérée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Image non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Récupérer une image par son ID (admin)",
        "tags": [
          "gallery"
        ]
      },
      "put": {
        "operationId": "GalleryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'image",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image mise à jour avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Image non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mettre à jour une image (admin)",
        "tags": [
          "gallery"
        ]
      },
      "delete": {
        "operationId": "GalleryController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de l'image",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image supprimée avec succès"
          },
          "401": {
            "description": "Non autorisé"
          },
          "404": {
            "description": "Image non trouvée"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer une image (admin)",
        "tags": [
          "gallery"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Aloha Secourisme",
    "description": "\n      API REST complète pour la gestion de l'application Aloha Secourisme.\n      \n      Cette API permet de gérer les formations, les inscriptions, les documents, \n      la galerie d'images et l'administration du système.\n      \n      ## Authentification\n      \n      La plupart des endpoints nécessitent une authentification par token JWT.\n      Pour vous authentifier, utilisez l'endpoint `/api/auth/login` pour obtenir un token,\n      puis incluez ce token dans l'en-tête d'autorisation de vos requêtes.\n      \n      ## Rôles\n      \n      - **Accès public** : Accessible sans authentification\n      - **Admin** : Nécessite d'être connecté avec un compte administrateur\n      - **Super Admin** : Nécessite d'être connecté avec un compte super administrateur\n      \n      ## Modules principaux\n      \n      - **Formations** : Gestion des sessions de formation\n      - **Inscriptions** : Gestion des inscriptions aux formations\n      - **Documents** : Gestion des documents téléchargeables\n      - **Galerie** : Gestion des images de la galerie\n      - **News** : Gestion des actualités\n      - **Admin/Users** : Gestion des utilisateurs administrateurs\n      - **Dashboard** : Statistiques et tableaux de bord\n      - **Settings** : Paramètres du système\n      - **Database** : Export/import de la base de données\n    ",
    "version": "1.0.0",
    "contact": {
...
    },
...
    }
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Entrez votre token JWT ici"
      }
    },
    "schemas": {
      "CreateFormationDto": {
        "type": "object",
        "properties": {
          "titre": {
            "type": "string",
            "example": "Formation PSC1",
            "description": "Titre de la formation"
          },
          "type": {
            "type": "string",
            "enum": [
              "PSC1",
              "PSE1",
              "PSE2",
              "BNSSA",
              "SSA",
              "SST",
              "BSB"
            ],
            "description": "Type de formation",
            "example": "PSC1"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-12-01T09:00:00Z",
            "description": "Date de la formation"
          },
          "duree": {
            "type": "string",
            "example": "7h",
            "description": "Durée de la formation"
          },
          "placesTotal": {
            "type": "number",
            "example": 12,
            "description": "Nombre total de places"
          },
          "prix": {
            "type": "number",
            "example": 89.99,
            "description": "Prix de la formation"
          },
          "lieu": {
            "type": "string",
            "example": "Lyon",
            "description": "Lieu de la formation"
          },
          "formateur": {
            "type": "string",
            "example": "Jean Dupont",
            "description": "Nom du formateur"
          }
        },
        "required": [
          "titre",
          "type",
          "date",
          "duree",
          "placesTotal",
          "prix",
          "lieu",
          "formateur"
        ]
      },
      "UpdateFormationDto": {
        "type": "object",
        "properties": {
          "titre": {
            "type": "string",
            "example": "Formation PSC1",
            "description": "Titre de la formation"
          },
          "type": {
            "type": "string",
            "enum": [
              "PSC1",
              "PSE1",
              "PSE2",
              "BNSSA",
              "SSA",
              "SST",
              "BSB"
            ],
            "description": "Type de formation",
            "example": "PSC1"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2023-12-01T09:00:00Z",
            "description": "Date de la formation"
          },
          "duree": {
            "type": "string",
            "example": "7h",
            "description": "Durée de la formation"
          },
          "placesTotal": {
            "type": "number",
            "example": 12,
            "description": "Nombre total de places"
          },
          "prix": {
            "type": "number",
            "example": 89.99,
            "description": "Prix de la formation"
          },
          "lieu": {
            "type": "string",
            "example": "Lyon",
            "description": "Lieu de la formation"
          },
          "formateur": {
            "type": "string",
            "example": "Jean Dupont",
            "description": "Nom du formateur"
          }
        }
      },
      "UpdateFormationStatusDto": {
        "type": "object",
        "properties": {
          "statut": {
            "type": "string",
            "enum": [
              "PLANIFIEE",
              "EN_COURS",
              "TERMINEE",
              "ANNULEE"
            ],
            "description": "Statut de la formation",
            "example": "EN_COURS"
          }
        },
        "required": [
          "statut"
        ]
      },
...
      "UpdateImageDto": {
        "type": "object",
        "properties": {
          "alt": {
            "type": "string",
            "example": "Formation PSC1 - juin 2023",
            "description": "Description de l'image"
          },
          "category": {
            "type": "string",
            "enum": [
              "formations",
              "activites",
              "evenements"
            ],
            "description": "Catégorie de l'image",
            "example": "formations"
          }
        }
      },
...
    }
  }
}